requirements/base.txt
Django==4.1.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djangorestframework-camel-case==1.3.0
drf-yasg==1.20.0
django-cors-headers==3.13.0
requests==2.28.1
drf-api-logger==1.1.12

python-dotenv==0.21.0
dj-database-url==1.0.0
mysqlclient==2.2.4

gunicorn==20.1.0
whitenoise==6.2.0
redis==4.3.2
celery==5.2.7
gevent==21.12.0
sendgrid==6.9.7
pusher==3.3.2
cloudinary==1.34.0
markdown==3.6